v2

+++ Data types/Letters/Sphere/Program.cs
@@ -10,7 +10,7 @@ namespace Sphere
     {
         static void Main(string[] args)
         {
-            int radius = Convert.ToInt32(Console.ReadLine());
+            double radius = Convert.ToInt32(Console.ReadLine());
             double volume = (4 * Math.PI * Math.Pow(radius, 3))/3;
             double surface = 4 * Math.PI * Math.Pow(radius, 2);
             Console.WriteLine($"the sphere has volume: {volume} and surface: {surface}");
